feat: 事前準備でフォルダ整理をした
