+++ test/mock.dart
@@ -1,4 +1,4 @@
-import 'package:github_searcher_flutter/model/client/github_repository_client.dart';
+import 'package:github_searcher_flutter/model/client/github_search_api_repository.dart';
 import 'package:mockito/annotations.dart';
 
 @GenerateMocks([
